+++ src/runtime.cpp
@@ -103,3 +103,4 @@ void ris_exit(int32_t code) {
 }
 
 } // extern "C"
+